feat(forum_conversation): make TopicNewsListView independant from TopicListView
